带有“不可腐烂”的lore的物品都不会腐烂 修复了高原生物群系和高原M不会下雨但也计算湿度的bug 修复了config设置不会腐烂仍然腐烂的问题 增加了箱子gui内食物腐烂 增加了一些配置在config.yml中：
#食物设置##########################################################
food:
  hunger:
    speed: 30   #每多少秒固定减少1点饥饿值
  rotten:
    invenable: true   #是否打开背包食物腐烂机制
    guienable: true   #是否打开箱子GUI内食物腐烂机制
    invspeed: 1   #背包食物每多少秒计算一次新鲜度-1
    guispeed: 1   #箱子GUI内食物每多少秒计算一次新鲜度-1

腐肉不再有新鲜度并修复了其在箱子里带有怪异lore的bug
修复多个食物腐烂只会变成一个腐烂食物的bug
冰箱投入使用：保质期延长10倍
修复一级机器仍需要机器的bug
插件支持热重载支持yum、plugman的重载
新增：合成树记忆化

击碎ute的方块会有击碎效果，只会掉落其部分合成材料
新增了可以在一次生命中重复选择角色的问题
修复了因为炎热而死亡报错的bug

人物技能与特性：

威尔逊-WILSON “我可是科学家”
属性
数值
最大理智值
200
初始最大血量
20
基础攻击倍率
x1.0
等级
初始0-最高0
普通：
--会长胡子，冬天起到保暖作用
--最大的优点是没有缺点

薇洛-WILLOW “燃烧是美丽的”
属性
数值
最大理智值
120
初始最大血量
20
基础攻击倍率
x1.0
等级
初始0-最高0
抑郁症：
--理智较低
纵火犯：
--靠近火、岩浆增加理智值
--免疫一切火焰伤害
--靠近水降低理智值，湿度对其影响更大

沃尔夫冈-WOLFGANG “看我的肌肉”
属性
数值
最大理智值
200
初始最大血量
见下
基础攻击倍率
见下
等级
0-4
胆小鬼-黑夜理智降低更快
大力士：
--血量上限=饱食度百分比*30
--攻击倍率=饱食度百分比x1.5

+++ src/HamsterYDS/UntilTheEnd/api/UntilTheEndApi.java
@@ -131,7 +131,7 @@ public class UntilTheEndApi {
         }
 
         public static String getSanityColor(Player player) {
-            int san = getValue(player, "san");
+            double san = getValue(player, "san");
             if (san >= 120) return HudProvider.yaml.getString("sanityColor.120");
             if (san >= 90) return HudProvider.yaml.getString("sanityColor.90");
             if (san >= 60) return HudProvider.yaml.getString("sanityColor.60");
@@ -141,7 +141,7 @@ public class UntilTheEndApi {
         }
 
         public static String getHumidityColor(Player player) {
-            int hum = getValue(player, "hum");
+            double hum = getValue(player, "hum");
             if (hum <= 5) return HudProvider.yaml.getString("humidityColor.5");
             if (hum <= 15) return HudProvider.yaml.getString("humidityColor.15");
             if (hum <= 25) return HudProvider.yaml.getString("humidityColor.25");
@@ -149,7 +149,7 @@ public class UntilTheEndApi {
         }
 
         public static String getTemperatureColor(Player player) {
-            int tem = getValue(player, "tem");
+            double tem = getValue(player, "tem");
             if (tem <= 15) return HudProvider.yaml.getString("temperatureColor.15");
             if (tem <= 50) return HudProvider.yaml.getString("temperatureColor.50");
             if (tem <= 75) return HudProvider.yaml.getString("temperatureColor.75");
@@ -157,7 +157,7 @@ public class UntilTheEndApi {
         }
         
         public static String getTirednessColor(Player player) {
-            int tir = getValue(player, "tir");
+            double tir = getValue(player, "tir");
             if (tir <= 25) return HudProvider.yaml.getString("tirednessColor.25");
             if (tir <= 50) return HudProvider.yaml.getString("tirednessColor.50");
             if (tir <= 75) return HudProvider.yaml.getString("tirednessColor.75");
@@ -165,8 +165,8 @@ public class UntilTheEndApi {
             return "";
         }
 
-        public static int getValue(Player player, String type) {
-            return (int) PlayerManager.check(player, type);
+        public static double getValue(Player player, String type) {
+            return PlayerManager.check(player, type);
         }
 
         public static void setValue(Player player, String type, int value) {
